Added testing files and CI/CD workflow file
I've added the basic test files for SpinStorm:
- test_routes.py
- test_models.py
- test_utils.py
- conftest.py
These files are found in the 'tests' folder.
In addition, I've added spinstorm-tests.yml to automatically run pytest when a pull request or push is done to main.
If there are any issues when pushing/pull requesting, let me know and I'll have a look.