RE:added integration and database tests into the suite
Had to revert the original changes since I didn't ignore the pycache folders.

The .gitignore I added in the root directory prevents every pycache folder in the repo from being committed.

The tests can be seen in /be/tests. The test folder needs to be in /be, as otherwise the tests have trouble importing the app's modules.

Finally, I added a testing config in be/config,py